MdePkg/PciSegmentLibSegmentInfo: Change parameters type
The current structure can lead to dependency on VC++ compiler support
functions when build for "IA32" with "/Od" switch, such as _aullshr().
So change the type of parameters in structure
PCI_SEGMENT_LIB_ADDRESS_STRUCTURE from UINT64 to UINT32.
https://bugzilla.tianocore.org/show_bug.cgi?id=1532

Cc: Michael D Kinney <michael.d.kinney@intel.com>
Cc: Liming Gao <liming.gao@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.1

+++ MdePkg/Library/PciSegmentLibSegmentInfo/PciSegmentLibCommon.c
@@ -17,13 +17,13 @@
 #include "PciSegmentLibCommon.h"
 
 typedef struct {
-  UINT64  Register : 12;
-  UINT64  Function : 3;
-  UINT64  Device : 5;
-  UINT64  Bus : 8;
-  UINT64  Reserved1 : 4;
-  UINT64  Segment : 16;
-  UINT64  Reserved2 : 16;
+  UINT32  Register : 12;
+  UINT32  Function : 3;
+  UINT32  Device : 5;
+  UINT32  Bus : 8;
+  UINT32  Reserved1 : 4;
+  UINT32  Segment : 16;
+  UINT32  Reserved2 : 16;
 } PCI_SEGMENT_LIB_ADDRESS_STRUCTURE;
 
 /**
